refactor: Move validateName utility to a dedicated file and update imports in NodeAddForm and NodeRenamingForm components
