+++ view/desktop/src/components/CollectionTree/utils/Form.ts
@@ -6,36 +6,6 @@ import {
   ItemConfigurationModel,
 } from "@repo/moss-collection";
 
-export const validateName = (
-  name: string,
-  restrictedNames: string[]
-): {
-  isValid: boolean;
-  message: string;
-} => {
-  if (!name) {
-    return {
-      isValid: false,
-      message: "The name cannot be empty",
-    };
-  }
-
-  const lowerCaseName = name.toLowerCase();
-  const lowerCaseRestrictedNames = restrictedNames.map((name) => name.toLowerCase());
-
-  if (lowerCaseRestrictedNames.includes(lowerCaseName)) {
-    return {
-      isValid: false,
-      message: `The "${name}" is already exists here`,
-    };
-  }
-
-  return {
-    isValid: true,
-    message: "",
-  };
-};
-
 export const createEntryKind = (
   name: string,
   path: string,
