+++ view/desktop/src/utils/index.ts
@@ -1,5 +1,7 @@
 export * from "./cn";
 export * from "./mergeRefs";
 export * from "./swapListByIndex";
+
 export * from "./swapListById";
 export * from "./swapObjectsById";
+export * from "./validateName";
